backend/requirements.txt
fastapi
uvicorn
pydantic-settings
pytest
httpx
sqlalchemy
psycopg2-binary
